refactor(core/linter-rules/a11y): convert core/a11y to linter rule
