+++ profiles/w3c.js
@@ -74,7 +74,7 @@ const modules = [
   import("../src/core/linter-rules/privsec-section.js"),
   import("../src/core/linter-rules/wpt-tests-exist.js"),
   import("../src/core/linter-rules/no-http-props.js"),
-  import("../src/core/a11y.js"),
+  import("../src/core/linter-rules/a11y.js"),
 ];
 
 async function domReady() {
